move losses
