+++ qhbmlib/inference/vqt_loss.py
